Admin stuff
Admin pages are all created and accessible.
There is now a fully functional user creation page that handles
everything needed and will be used as a model for the other admin pages.

Some UI tweaks.
Add a (blank) user preference page that is accessible from next to the
log out button.